Remove `Base.copyDir()`
There was already a nearly identical `FileUtils.copy()` that copies
directories recursively.

The only difference is that now hidden files *are* copied, but version
control files (according the list in FileUtils) are not. Since this only
affects the copying of the "data" directory during save as, this
should not be much of a problem.