Moving "Edit" menu actions from Editor to EditorTab
Before the Edit menu was build in Editor and shared between all
editor tabs. The various ActionListeners were shared as well and
they forwarded events to the current active tab using the
Editor.getCurrentTab() method.

This commit move the menu creation responsibility into the EditorTabs
that now have a method:

  abstract public JMenuItem[] getEditMenuEntries();

It's now the Editor that adds and remove the relevant menus every
time tabs are swapped.