Simplified ContributedLibrary class hierarchy
This commit aims to remove all the subclasses of ContributedLibrary
and simplify the code that handles the various subclasses UserLibrary
and LegacyUserLibrary.

After the refactoring the subclasses are reduced to, basically, empty
factories. Since the class RTTI is never used the static factory
method may be moved to the top level ContributedLibrary class and the
subclasses removed altogether.